requirements.txt
streamlit
pinecone-client
ipoly
